% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb.R
\name{dr_default_priors}
\alias{dr_default_priors}
\title{Default hierarchical priors}
\usage{
dr_default_priors(dim_phi)
}
\arguments{
\item{dim_phi}{dimension of the individual-parameter vector}
}
\description{
Default hierarchical priors
}
