% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nash.R
\name{dr_demand_constant}
\alias{dr_demand_constant}
\title{Constant-incidence heuristic demand (MNL shares x fixed incidence)}
\usage{
dr_demand_constant(shelf_fn, p, BigB, incidence)
}
\arguments{
\item{shelf_fn}{function(p) returning the shelf design matrix (J x P) at
price vector p}

\item{p}{price vector}

\item{BigB}{stacked coefficient matrix}

\item{incidence}{fixed purchase incidence applied to every task}
}
\description{
Constant-incidence heuristic demand (MNL shares x fixed incidence)
}
