% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{dr_draw_betas}
\alias{dr_draw_betas}
\title{Draw individual coefficients from a multivariate-normal population}
\usage{
dr_draw_betas(N, beta_bar, Sigma, seed)
}
\arguments{
\item{N}{respondents}

\item{beta_bar}{population mean}

\item{Sigma}{population covariance}

\item{seed}{RNG seed}
}
\description{
Draw individual coefficients from a multivariate-normal population
}
