% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{dr_laplace_lmd}
\alias{dr_laplace_lmd}
\title{Laplace-approximation log marginal likelihood for the aggregate model}
\usage{
dr_laplace_lmd(fit_agg)
}
\arguments{
\item{fit_agg}{output of \code{\link[=dr_bayes_agg]{dr_bayes_agg()}}}
}
\description{
Laplace-approximation log marginal likelihood for the aggregate model
}
