% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{dr_optimal_cuts}
\alias{dr_optimal_cuts}
\title{Information-optimal cut points}
\usage{
dr_optimal_cuts(W, model, m0, s)
}
\arguments{
\item{W}{scale points}

\item{model}{link}

\item{m0, s}{inclusive-value distribution}
}
\value{
list(cut, omega)
}
\description{
Maximizes the mean second-stage information weight over the inclusive-value
distribution.
}
